+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/predicate/PredicateInitializer.java
@@ -2,6 +2,7 @@ package io.github.sakurawald.fuji.module.initializer.predicate;
 
 import io.github.sakurawald.fuji.core.auxiliary.minecraft.InventoryHelper;
 import io.github.sakurawald.fuji.core.auxiliary.minecraft.PlayerHelper;
+import io.github.sakurawald.fuji.core.auxiliary.minecraft.TextHelper;
 import io.github.sakurawald.fuji.core.document.annotation.ColorBox;
 import io.github.sakurawald.fuji.core.document.annotation.Document;
 import io.github.sakurawald.fuji.core.auxiliary.minecraft.CommandHelper;
@@ -14,10 +15,13 @@ import io.github.sakurawald.fuji.core.command.argument.wrapper.impl.Dimension;
 import io.github.sakurawald.fuji.core.command.argument.wrapper.impl.GreedyString;
 import io.github.sakurawald.fuji.module.initializer.ModuleInitializer;
 import io.github.sakurawald.fuji.core.document.descriptor.PermissionDescriptor;
+import java.util.Objects;
+import java.util.function.BiPredicate;
 import net.minecraft.command.argument.ItemPredicateArgumentType;
 import net.minecraft.item.ItemStack;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.text.Text;
 import net.minecraft.world.GameMode;
 
 import java.util.Optional;
@@ -28,7 +32,7 @@ import java.util.Optional;
 
 @ColorBox(id = 1751978705157L, color = ColorBox.ColorBlockTypes.NOTE, value = """
     ◉ How it works?
-    A `predicate command` will returns `SUCCESS` if `true`, and `FAILED` if `false`.
+    A `predicate command` will return an `integer value` to represent the `SUCCESS` or `FAILED` test result.
     It relies on the `return value of command` in vanilla Minecraft.
     See https://minecraft.fandom.com/wiki/Commands/return
 
@@ -37,6 +41,22 @@ import java.util.Optional;
 @ColorBox(id = 1751978789271L, color = ColorBox.ColorBlockTypes.EXAMPLE, value = """
     ◉ Test a condition, and then run a command if success.
     Issue: `/run as player Alice chain has-perm? %player:name% 4 chain say value is true`
+    """)
+@ColorBox(id = 1754448424606L, color = ColorBox.ColorBlockTypes.EXAMPLE, value = """
+    ◉ Test if a player is holding an apple
+    Issue: `/command-debug is-holding? Steve minecraft:apple`
+
+    ◉ Test if a player has specified string permission
+    Issue: `/command-debug has-perm? Steve fuji.permission.back`
+
+    ◉ Test if a player's health >= 10
+    Issue: `/command-debug <=? Steve 10 %player:health%`
+
+    ◉ Test if a player killed more than 100 zombies.
+    Issue: `/command-debug <=? Steve 100 %player:statistic_raw minecraft:deaths%`
+
+    ◉ More examples
+    You can see more examples in `command_meta.IF` module and `command_meta.chain` module.
     """)
 
 
@@ -112,5 +132,49 @@ public class PredicateInitializer extends ModuleInitializer {
     private static int $isInGameMode(@CommandSource ServerCommandSource source, ServerPlayerEntity player, GameMode gameMode) {
         boolean value = player.interactionManager.getGameMode().equals(gameMode);
         return CommandHelper.Return.returnBoolean(source, value);
+    }
+
+    @CommandNode("=?")
+    private static int $equalsNumber(@CommandSource ServerCommandSource source, ServerPlayerEntity player, double value, GreedyString numericValueProvider) {
+        return compareNumericValue(source, player, value, numericValueProvider, Objects::equals);
+    }
+
+    @CommandNode("!=?")
+    private static int $notEqualsNumber(@CommandSource ServerCommandSource source, ServerPlayerEntity player, double value, GreedyString numericValueProvider) {
+        return compareNumericValue(source, player, value, numericValueProvider, (a, b) -> !Objects.equals(a, b));
+    }
+
+    @CommandNode(">?")
+    private static int $greaterThanNumber(@CommandSource ServerCommandSource source, ServerPlayerEntity player, double value, GreedyString numericValueProvider) {
+        return compareNumericValue(source, player, value, numericValueProvider, (a, b) -> a > b);
+    }
+
+    @CommandNode(">=?")
+    private static int $greaterEqualsThanNumber(@CommandSource ServerCommandSource source, ServerPlayerEntity player, double value, GreedyString numericValueProvider) {
+        return compareNumericValue(source, player, value, numericValueProvider, (a, b) -> a >= b);
+    }
+
+    @CommandNode("<?")
+    private static int $lessThanNumber(@CommandSource ServerCommandSource source, ServerPlayerEntity player, double value, GreedyString numericValueProvider) {
+        return compareNumericValue(source, player, value, numericValueProvider, (a, b) -> a < b);
+    }
+
+    @CommandNode("<=?")
+    private static int $lessEqualsThanNumber(@CommandSource ServerCommandSource source, ServerPlayerEntity player, double value, GreedyString numericValueProvider) {
+        return compareNumericValue(source, player, value, numericValueProvider, (a, b) -> a <= b);
+    }
+
+    private static int compareNumericValue(ServerCommandSource source, ServerPlayerEntity player, double value, GreedyString numericValueProvider, BiPredicate<Double, Double> predicate) {
+        String $numericValueProvider = numericValueProvider.getValue();
+        Text numericValueText = TextHelper.getTextByValue(player, $numericValueProvider);
+        String numericValueString = TextHelper.Operators.visitString(numericValueText);
+        try {
+            double numericValue = Double.parseDouble(numericValueString);
+            boolean testResult = predicate.test(value, numericValue);
+            return CommandHelper.Return.returnBoolean(source, testResult);
+        } catch (NumberFormatException e) {
+            TextHelper.sendTextByKey(source, "placeholder.number.parse.failed", TextHelper.Parsers.escapeTags($numericValueProvider));
+            return CommandHelper.Return.FAIL;
+        }
     }
 }